requirements-dev.txt
-r requirements.txt
pytest==2.6.4
pytest-django==2.8.0
